Moved func 'RenderByCommand()' to renderer.go
