--- pkg/diff/diffbycommand.go
@@ -1,81 +1,0 @@
-// Copyright 2024 The PipeCD Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package diff
-
-import (
-	"bytes"
-	"fmt"
-	"os"
-	"os/exec"
-
-	"sigs.k8s.io/yaml"
-)
-
-// RenderByCommand converts the given objects into yaml and then runs the command to compare them.
-func RenderByCommand(command string, old, new interface{}) ([]byte, error) {
-	_, err := exec.LookPath(command)
-	if err != nil {
-		return nil, fmt.Errorf("failed to find the command '%s', err = %w", command, err)
-	}
-
-	oldBytes, err := yaml.Marshal(old)
-	if err != nil {
-		return nil, err
-	}
-	newBytes, err := yaml.Marshal(new)
-	if err != nil {
-		return nil, err
-	}
-
-	oldFile, err := os.CreateTemp("", "old")
-	if err != nil {
-		return nil, err
-	}
-	defer os.Remove(oldFile.Name())
-	if _, err := oldFile.Write(oldBytes); err != nil {
-		return nil, err
-	}
-
-	newFile, err := os.CreateTemp("", "new")
-	if err != nil {
-		return nil, err
-	}
-	defer os.Remove(newFile.Name())
-	if _, err := newFile.Write(newBytes); err != nil {
-		return nil, err
-	}
-
-	var stdout, stderr bytes.Buffer
-	cmd := exec.Command(command, "-u", "-N", oldFile.Name(), newFile.Name())
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-	err = cmd.Run()
-	if stdout.Len() > 0 {
-		// diff exits with a non-zero status when the files don't match.
-		// Ignore that failure as long as we get output.
-		err = nil
-	}
-	if err != nil {
-		return nil, fmt.Errorf("failed to run diff, err = %w, %s", err, stderr.String())
-	}
-
-	// Remove two-line header from output.
-	data := bytes.TrimSpace(stdout.Bytes())
-	rows := bytes.SplitN(data, []byte("\n"), 3)
-	if len(rows) == 3 {
-		return rows[2], nil
-	}
-	return data, nil
-}
